Practice

+++ Java/RoomRental.java
@@ -29,8 +29,9 @@ public class RoomRental {
             if(type.getText().equalsIgnoreCase("A")){
                 if(Integer.parseInt(numofdays.getText()) > 12){
                     JOptionPane.showMessageDialog(null, 
-                    "Your rental costs " + 0.9*(450000*Integer.parseInt(numofdays.getText())) 
-                    + " VND","Java Room Rental", JOptionPane.INFORMATION_MESSAGE);
+                    "Your rental costs " + (int) 
+                    (0.9*(450000*Integer.parseInt(numofdays.getText()))) + " VND",
+                    "Java Room Rental", JOptionPane.INFORMATION_MESSAGE);
                     return;
                 }
                 JOptionPane.showMessageDialog(null, 
@@ -40,8 +41,9 @@ public class RoomRental {
             if(type.getText().equalsIgnoreCase("B")){
                 if(Integer.parseInt(numofdays.getText()) > 12){
                     JOptionPane.showMessageDialog(null, 
-                    "Your rental costs " + 0.92*(350000*Integer.parseInt(numofdays.getText())) 
-                    + " VND","Java Room Rental", JOptionPane.INFORMATION_MESSAGE);
+                    "Your rental costs " + (int)
+                    (0.92*(350000*Integer.parseInt(numofdays.getText()))) + " VND",
+                    "Java Room Rental", JOptionPane.INFORMATION_MESSAGE);
                     return;
                 }
                 JOptionPane.showMessageDialog(null, 
@@ -51,8 +53,9 @@ public class RoomRental {
             if(type.getText().equalsIgnoreCase("C")){
                 if(Integer.parseInt(numofdays.getText()) > 12){
                     JOptionPane.showMessageDialog(null, 
-                    "Your rental costs " + 0.92*(25000*Integer.parseInt(numofdays.getText())) 
-                    + " VND","Java Room Rental", JOptionPane.INFORMATION_MESSAGE);
+                    "Your rental costs " + (int) 
+                    (0.92*(25000*Integer.parseInt(numofdays.getText()))) + " VND",
+                    "Java Room Rental", JOptionPane.INFORMATION_MESSAGE);
                     return;
                 }
                 JOptionPane.showMessageDialog(null,
